reference_implementations/gcp/step6/vertex/requirements.txt
google-cloud-aiplatform[prediction]==1.55.0
torch==2.2.0
transformers==4.38.1
accelerate==0.27.0
cloud-sql-python-connector==1.9.2
SQLAlchemy==2.0.30
pg8000==1.31.2
pandas==2.0.3

numpy<2
